+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+## Posh 0.1.2:
+
+* New Demos
+  * Posh.demo.ps1 (#95)
+  * Posh.Prompt.demo.ps1 (#96)
+  * Posh Profile Customization (#97)
+* New formatters:
+  * Adding Formatting for Match (#100)
+  * Adding Variable Formatting (#102)  
+* Extending Modules
+  * Adding PSModuleInfo.Test(s) (#104)
+* Prompt Customization:
+  * Posh.Prompt.Undo() (#103)
+* Profile Customization:
+  * Posh.Profiles.RemoveModule (#105)
+  * Posh.Profiles.New() (#101)
+  * Posh.Profile.ImportModule (#93)
+  * Posh.Profile.Matches (#99)
+  * Posh.Profile.Replace (#92)
+  * Posh.Profile.get_File (#76)
+
+---
+
 ## Posh 0.1.1:
 
 * Posh can be sponsored (please show your support) (#77)